+++ Sources/TuistAcceptanceTesting/ServerAcceptanceTestCase.swift
@@ -1,0 +1,30 @@
+import Foundation
+import XCTest
+
+@testable import TuistKit
+
+open class ServerAcceptanceTestCase: TuistAcceptanceTestCase {
+    public var fullHandle: String = ""
+    public var organizationHandle: String = ""
+    public var projectHandle: String = ""
+
+    override open func setUp() async throws {
+        try await super.setUp()
+        try setUpFixture(.iosAppWithFrameworks)
+        organizationHandle = String(UUID().uuidString.prefix(12).lowercased())
+        projectHandle = String(UUID().uuidString.prefix(12).lowercased())
+        fullHandle = "\(organizationHandle)/\(projectHandle)"
+        let email = try XCTUnwrap(ProcessInfo.processInfo.environment[EnvKey.authEmail.rawValue])
+        let password = try XCTUnwrap(ProcessInfo.processInfo.environment[EnvKey.authPassword.rawValue])
+        try await run(AuthCommand.self, "--email", email, "--password", password)
+        try await run(OrganizationCreateCommand.self, organizationHandle)
+        try await run(ProjectCreateCommand.self, fullHandle)
+    }
+
+    override open func tearDown() async throws {
+        try await run(ProjectDeleteCommand.self, fullHandle)
+        try await run(OrganizationDeleteCommand.self, organizationHandle)
+        try run(LogoutCommand.self)
+        try await super.tearDown()
+    }
+}
